LayoutTests/svg/custom/xlink-prefix-generation-in-attributes-expected.txt

<svg xmlns="http://www.w3.org/2000/svg" NS1:href="#foo" xmlns:NS1="http://www.w3.org/1999/xlink"/>
<svg xmlns="http://www.w3.org/2000/svg" NS1:href="#foo" xmlns:NS1="http://www.w3.org/1999/xlink"/>
<svg xmlns="http://www.w3.org/2000/svg" xlink:href="#foo" xmlns:xlink="http://www.w3.org/1999/xlink"/>
<svg xmlns="http://www.w3.org/2000/svg" xlink:href="#foo" xmlns:xlink="http://www.w3.org/1999/xlink"/>
